Add experimental proplist.txt support
This is a file automatically generated by the Dark engine via the
:dump_props_full command. It's fairly straightforward to parse and
describes the data format of all of the properties, links, effects,
and propagators understood by the engine; it even gives you enum and
bitfield names!

Unfortunately, for aggregate data types, it is not guaranteed to output
the fields in the same order they appear on disk, so it can only be
safely used automatically for non-aggregate types; anything else
requires human intervention to make sure the field ordering is correct.