Split DB objects up by type
It turns out that IDs are not unique across types, i.e. the ID keyspace
for brushes is distinct from the one for in-map entities. :( So we split
it up and require the caller to specify a type. If they don't, reads
default to 'entity' and everything else throws.

Also adds some documentation to the db library.