Fix db:clone()
Since each object's meta.db field contained a reference to its
containing db, and cloning the db dropped the metatables on this,
anything that used meta.db would fail. With this we instead store the db
in the object's metatable, which means it gets properly (re)instated
when we access the object through db object lookup methods.